IGNITE-17267 wip

+++ modules/page-memory/src/main/java/org/apache/ignite/internal/pagememory/persistence/checkpoint/Checkpoint.java
@@ -28,7 +28,7 @@ class Checkpoint {
     final CheckpointProgressImpl progress;
 
     /** Number of dirty pages. */
-    final int dirtyPagesSize;
+    final long dirtyPagesSize;
 
     /**
      * Constructor.
