+++ modules/page-memory/src/main/java/org/apache/ignite/internal/pagememory/persistence/checkpoint/CheckpointProgress.java
@@ -48,5 +48,5 @@ public interface CheckpointProgress {
     /**
      * Returns number of dirty pages in current checkpoint. If checkpoint is not running, returns {@code 0}.
      */
-    int currentCheckpointPagesCount();
+    long currentCheckpointPagesCount();
 }
